Minor edits.
On Windows Store, if you have never downloaded an app, the install button will be Get. If you have installed the app in the past and then uninstalled it,
the text on the install button is Install. The text in these two pages was updated to say to click the Get or the Install button.

Also changed your to you in one place.